refactor ContactPage helper logic - move react related logic to utils directory in the ContactPage components directory
