+++ src/lib/helpers/contact.tsx
@@ -1,5 +1,3 @@
-import { ClonedElement, Email, IconWithContent, Link, Typography } from '@components';
-import { ReactNode } from 'react';
 import { IconType } from 'react-icons';
 import { FaEnvelope, FaGithub, FaLinkedin, FaMapMarkerAlt } from 'react-icons/fa';
 
@@ -10,6 +8,21 @@ type ContactData = {
   email: string;
   githubLink: string;
   linkedInLink: string;
+};
+
+export enum ContactListType {
+  Location = 'location',
+  Email = 'email',
+  Github = 'github',
+  LinkedIn = 'linkedin',
+}
+
+export type ContactList = {
+  type: ContactListType;
+  icon: IconType;
+  text: string;
+  url?: string;
+  title?: string;
 };
 
 export const getContactData = ({ location, email, githubLink, linkedInLink }: ContactData) => {
@@ -37,49 +50,4 @@ export const getContactData = ({ location, email, githubLink, linkedInLink }: Co
       text: removeHttpeProtocol(linkedInLink),
     },
   ];
-};
-
-enum ContactListType {
-  Location = 'location',
-  Email = 'email',
-  Github = 'github',
-  LinkedIn = 'linkedin',
-}
-
-type ContactList = {
-  type: ContactListType;
-  icon: IconType;
-  text: string;
-  url?: string;
-  title?: string;
-};
-
-export const generateContactListItems = (contactData: ContactList[]): ReactNode[] => {
-  const textComponent = <Typography variant="heading5" />;
-
-  return contactData.map((contactItem, index) => {
-    const { icon: Icon, title, url, text, type } = contactItem;
-
-    const renderListItem = () => {
-      if (!url || !title) {
-        return <ClonedElement element={textComponent}>{text}</ClonedElement>;
-      }
-
-      if (type === ContactListType.Email) {
-        return <Email href={url} text={text} title={title} childComponent={textComponent} />;
-      }
-
-      return (
-        <Link href={url} title={title} target="_blank">
-          <ClonedElement element={textComponent}>{text}</ClonedElement>
-        </Link>
-      );
-    };
-
-    return (
-      <IconWithContent key={`contact-item-${type}-${index}`} icon={<Icon fontSize={20} />}>
-        {renderListItem()}
-      </IconWithContent>
-    );
-  });
 };
